% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumcommits.R
\name{cum_commits}
\alias{cum_commits}
\title{Cummulative commits plot}
\usage{
cum_commits(x, exclude = c("sckott", "karthik", "cboettig", "jeroenooms"),
  exclude_core = FALSE)
}
\arguments{
\item{x}{File name, e.g., github_commits_2016-05-19.csv. Required.}

\item{exclude}{(character) github user names to exclude}

\item{exclude_core}{(logical) exclude core user names. Default: \code{FALSE}}
}
\description{
Cummulative commits plot
}
\examples{
\dontrun{
cum_commits(x = "github_commits_2016-05-19.csv")
cum_commits(x = "github_commits_2016-05-19.csv", exclude_core = TRUE)
}
}

