% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_downloads.R
\name{cum_downloads}
\alias{cum_downloads}
\title{Cummulative downloads plot}
\usage{
cum_downloads(x = NULL, top_n = 10)
}
\arguments{
\item{x}{File name, e.g., cran_downloads_2016-05-19.csv. Default is \code{NULL},
and if so, we look for file with most recent date in its file name in
\code{rappdirs::user_cache_dir("rostats")} + "/cran_downloads/"}

\item{top_n}{(numeric/integer) number of packages to plot data for,
starting from the most downloaded}
}
\description{
Cummulative downloads plot
}
\examples{
\dontrun{
cum_downloads()
}
}

