% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathercommits.R
\name{gather_commits}
\alias{gather_commits}
\title{Get rOpenSci commits}
\usage{
gather_commits(x, file = Sys.Date(), token = NULL)
}
\arguments{
\item{x}{One or more R packages, of the form \code{ropensci/rfishbase},
case sensitive}

\item{file}{A date, or other character string to make the file
ending unique. Default: \code{Sys.Date()}}

\item{token}{A GitHub Personal Access Token (PAT). If not given,
we look in your env vars for GITHUB_PAT}
}
\value{
writes a data.frame to \code{rappdirs::user_cache_dir("rostats")},
prints message of path to file, and returns file path itself
}
\description{
Get rOpenSci commits
}
\examples{
\dontrun{
library("dplyr")
library("ropkgs")
library("rostats")
pkgs <- ropensci_pkgs()


library("dplyr")
url <- "https://raw.githubusercontent.com/ropensci/roregistry/gh-pages/registry.json"
df <- jsonlite::fromJSON(url)
(pkgs <- tbl_df(df$packages) \%>\%
  # filter(on_cran, !cran_archived) \%>\%
  # filter(name != "pleiades") \%>\%
  .$url \%>\%
  gsub('https://github.com/', '', .)
)

pkgs[3] <- "GuangchuangYu/treeio"

z <- gather_commits(x = pkgs)
# gather_commits(x = pkgs$owner_repo)
}
}
