% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_data.R
\name{gather_cran}
\alias{gather_cran}
\alias{gather_crans}
\alias{cran_first_date}
\title{Gather CRAN details}
\usage{
gather_cran(pkg)

gather_crans(pkg)

cran_first_date(x)
}
\arguments{
\item{pkg}{(character) a package name}

\item{x}{a data.frame}
}
\description{
Gather CRAN details
}
\details{
While \code{\link{gather_commits}} and \code{\link{gather_downloads}} cache
data because data can take a long time to download, this function shouldn't take that
long, even with hundreds of of pkgs.
}
\examples{
\dontrun{
library("rostats")
library("ropkgs")
library("dplyr")
url <- "https://raw.githubusercontent.com/ropensci/roregistry/gh-pages/registry.json"
df <- jsonlite::fromJSON(url)
(pkgs <- tbl_df(df$packages) \%>\%
  filter(on_cran | on_bioc) \%>\%
  .$name)
sort(pkgs)

# get first date on CRAN for each pkg
res <- gather_crans(pkgs)
alldat <- dplyr::bind_rows(res)

# get new packages on CRAN, arranged by date
alldat \%>\% cran_first_date() \%>\% arrange(desc(date)) \%>\% top_n(10) \%>\%
  readr::write_csv("../biweekly/data/newpkgs.csv")

# get new packages & new versions on CRAN, arranged by date
alldat \%>\% arrange(desc(date)) \%>\% filter(date > '2019-04-14') \%>\% top_n(30) \%>\%
  readr::write_csv("../biweekly/data/newversions.csv")
}
}
