% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_downloads.R
\name{gather_downloads}
\alias{gather_downloads}
\title{Get rOpenSci commits}
\usage{
gather_downloads(x, from = NULL, to = Sys.Date() - 1, file = Sys.Date())
}
\arguments{
\item{x}{One or more R packages, of the form \code{ropensci/rfishbase},
case sensitive}

\item{from}{(Date) A date to start from. Default is \code{NULL}, and we
get the first date published on CRAN to start from using \code{\link{gather_cran}}
and \code{\link{cran_first_date}}}

\item{to}{(Date) A date to end at. Default: \code{Sys.Date() - 1}}

\item{file}{A date, or other character string to make the file
ending unique. Default: \code{Sys.Date()}}
}
\value{
writes a data.frame to \code{rappdirs::user_cache_dir("rostats")} +
'/cran_downloads/', prints message of path to file, and returns file path itself
}
\description{
Get rOpenSci commits
}
\examples{
\dontrun{
library("ropkgs")
library("dplyr")
df <- ro_pkgs()
pkgs <- tbl_df(df$packages) \%>\%
  filter(on_cran) \%>\%
  .$name \%>\%
  sort
pkgs <- sub("redland-bindings", "redland", pkgs)
gather_downloads(x = pkgs[1])
# gather_downloads(x = pkgs)
}
}
