% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_cran}
\alias{get_cran}
\alias{get_github}
\title{Read data from file path in package}
\usage{
get_cran(x)

get_github(x)
}
\arguments{
\item{x}{File name, e.g., github_commits_2016-05-19.csv. Required.}
}
\value{
a data.frame
}
\description{
Read data from file path in package
}
\details{
file must be in \code{extdata}
}

