% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropensci_pkgs.R
\name{ropensci_pkgs}
\alias{ropensci_pkgs}
\title{Get rOpenSci packages, actual R packages, via the ropensci registry}
\usage{
ropensci_pkgs(only_on_cran = FALSE)
}
\arguments{
\item{only_on_cran}{(logical). return only pkgs on CRAN. Default: \code{FALSE}}
}
\description{
Get rOpenSci packages, actual R packages, via the ropensci registry
}
\examples{
\dontrun{
ropensci_pkgs()
ropensci_pkgs(TRUE)
}
}

