% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropensci_review_dates.R
\name{ropensci_review_dates}
\alias{ropensci_review_dates}
\title{get date that submissions were approved in ropensci software review repo}
\usage{
ropensci_review_dates(token = NULL)
}
\arguments{
\item{token}{(optional) a github PAT}
}
\description{
get date that submissions were approved in ropensci software review repo
}
\examples{
library(dplyr)
# all approved packages
df <- ropensci_review_dates()
# filter to some subset
## those approved in 2018
tbl_df(df) \%>\%
  arrange(date_approved) \%>\%
  filter(date_approved >= "2018-01-01", date_approved < "2019-01-01") \%>\%
  data.frame \%>\%
  NROW
}
