% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumcontribs.R
\name{cum_contribs}
\alias{cum_contribs}
\title{Cummulative commits plot}
\usage{
cum_contribs(x = NULL, exclude = c("sckott", "karthik", "cboettig",
  "jeroen"), exclude_core = FALSE)
}
\arguments{
\item{x}{File name, e.g., github_commits_2016-05-19.csv. Default is \code{NULL},
and if so, we look for file with most recent date in its file name in
\code{rappdirs::user_cache_dir("rostats")} + "/commits/"}

\item{exclude}{(character) github user names to exclude}

\item{exclude_core}{(logical) exclude core user names. Default: \code{FALSE}}
}
\description{
Cummulative commits plot
}
\examples{
\dontrun{
cum_contribs()
cum_contribs(exclude_core = TRUE)
}
}

