% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_downloads.R
\name{cum_downloads}
\alias{cum_downloads}
\title{Cummulative downloads plot}
\usage{
cum_downloads(x, top_n = 10)
}
\arguments{
\item{x}{File name, e.g., cran_downloads_2016-05-20.csv. Required.}

\item{top_n}{(numeric/integer) number of packages to plot data for,
starting from the most downloaded}
}
\description{
Cummulative downloads plot
}
\examples{
\dontrun{
cum_downloads(x = "cran_downloads_2016-05-23.csv")
}
}

