% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_data.R
\name{gather_cran}
\alias{gather_cran}
\alias{gather_crans}
\alias{cran_first_date}
\title{Gather CRAN details}
\usage{
gather_cran(pkg)

gather_crans(pkg)

cran_first_date(x)
}
\arguments{
\item{pkg}{(character) a package name}
}
\description{
Gather CRAN details
}
\details{
While \code{\link{gather_commits}} and \code{\link{gather_downloads}} cache
data because data can take a long time to download, this function shouldn't take that
long, even with hundreds of of pkgs.
}
\examples{
\dontrun{
gather_cran("taxize")

library("ropkgs")
library("dplyr")
#df <- ro_pkgs()
url <- "https://raw.githubusercontent.com/ropensci/roregistry/gh-pages/registry.json"
df <- jsonlite::fromJSON(url)
(pkgs <- tbl_df(df$packages) \%>\%
  filter(on_cran, !cran_archived) \%>\%
  filter(name != "pleiades") \%>\%
  # filter(on_cran | on_bioc) \%>\%
  .$name)

gather_cran(pkgs[1])

# get first date on CRAN for each pkg
res <- gather_crans(pkgs)
alldat <- dplyr::bind_rows(res)

# gather_crans(pkgs$name)

# get new packages on CRAN, arranged by date
alldat \%>\% cran_first_date() \%>\% arrange(desc(date)) \%>\% top_n(10)

# get new packages & new versions on CRAN, arranged by date
alldat \%>\% arrange(desc(date)) \%>\% filter(date > '2017-06-15') \%>\% top_n(30)
}
}
