% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_data.R
\name{gather_cran}
\alias{cran_first_date}
\alias{gather_cran}
\alias{gather_crans}
\title{Gather CRAN details}
\usage{
gather_cran(pkg)

gather_crans(pkg)

cran_first_date(x)
}
\arguments{
\item{pkg}{(character) a package name}
}
\description{
Gather CRAN details
}
\examples{
\dontrun{
gather_cran("taxize")

library("ropkgs")
library("dplyr")
df <- ro_pkgs()
pkgs <- df$packages \%>\%
  filter(on_cran) \%>\%
  .$name

gather_cran(pkgs[1])

res <- gather_crans(pkgs)
alldat <- dplyr::bind_rows(res)
cran_first_date(alldat)

# gather_crans(pkgs$name)
}
}

